requirements.txt
# Core dependencies
pandas
matplotlib
certifi
python-dotenv

# API client for market data
polygon-api-client

# Optional: for plotting inside notebooks
jupyterlab

# If you plan to use machine learning later
scikit-learn
xgboost
